Reorganise the source code
