+++ proof_of_concept/components/__init__.py
@@ -1,0 +1,1 @@
+"""Main components of sites and registry."""
