+++ proof_of_concept/definitions/__init__.py
@@ -1,0 +1,1 @@
+"""General definitions used throughout the code."""
