+++ proof_of_concept/definitions/policy.py
@@ -1,0 +1,7 @@
+"""Classes for describing and managing policies."""
+from proof_of_concept.definitions.signable import Signable
+
+
+class Rule(Signable):
+    """Abstract base class for policy rules."""
+    pass
