+++ proof_of_concept/policy/replication.py
@@ -1,0 +1,47 @@
+"""Support for replication of policies."""
+from proof_of_concept.definitions.policy import Rule
+from proof_of_concept.policy.definitions import PolicyUpdate
+from proof_of_concept.policy.rules import (
+        InAssetCollection, InPartyCollection, MayAccess, ResultOfIn,
+        ResultOfDataIn, ResultOfComputeIn)
+from proof_of_concept.replication import CanonicalStore, ObjectValidator
+
+from cryptography.hazmat.primitives.asymmetric.rsa import RSAPublicKey
+
+
+class RuleValidator(ObjectValidator[Rule]):
+    """Validates incoming policy rules by checking signatures."""
+    def __init__(self, namespace: str, key: RSAPublicKey) -> None:
+        """Create a RuleValidator.
+
+        Checks that rules apply to the given namespace, and that they
+        have been signed by the owner of that namespace.
+
+        Args:
+            namespace: The namespace to expect rules for.
+            key: The key to validate the rules with.
+        """
+        self._namespace = namespace
+        self._key = key
+
+    def is_valid(self, rule: Rule) -> bool:
+        """Return True iff the rule is properly signed."""
+        if isinstance(rule, ResultOfDataIn):
+            namespace = rule.data_asset[3:].split('.')[0]
+        elif isinstance(rule, ResultOfComputeIn):
+            namespace = rule.compute_asset[3:].split('.')[0]
+        elif isinstance(rule, MayAccess):
+            namespace = rule.asset[3:].split('.')[0]
+        elif isinstance(rule, InAssetCollection):
+            namespace = rule.asset[3:].split('.')[0]
+        elif isinstance(rule, InPartyCollection):
+            namespace = rule.collection[3:].split('.')[0]
+
+        if namespace != self._namespace:
+            return False
+        return rule.has_valid_signature(self._key)
+
+
+class PolicyStore(CanonicalStore[Rule]):
+    """A canonical store for policy rules."""
+    UpdateType = PolicyUpdate
