+++ proof_of_concept/registry/__init__.py
@@ -1,0 +1,1 @@
+"""The central registry and its replication mechanism."""
