+++ proof_of_concept/rest/__init__.py
@@ -1,0 +1,1 @@
+"""REST APIs, servers and clients for the DDM."""
