+++ proof_of_concept/rest/definitions.py
@@ -1,0 +1,5 @@
+"""General definitions for REST APIs."""
+from typing import Any, Dict
+
+
+JSON = Dict[str, Any]
