+++ proof_of_concept/rest/registry_api.yaml
